requirements.txt
Flask
Flask-SocketIO
deepgram-sdk
python-dotenv
librosa
nltk
numpy
anthropic
SQLAlchemy
pytest
scipy
soxr
pydub
soundfile
spacy
thinc
PySoundFile 
sndfile